+++ backend/internal/models/posts.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type Post struct {
-	ID        int       `json:"id"`
-	UserID    int       `json:"userId"`
+	ID        int64     `json:"id"`
+	UserID    int64     `json:"userId"`
 	Content   string    `json:"content"`
 	CreatedAt time.Time `json:"createdAt,omitempty"`
 	UpdatedAt time.Time `json:"updatedAt"`
